% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_block.R
\name{available_engines}
\alias{available_engines}
\title{Available engines for \code{new_price_block()}}
\usage{
available_engines()
}
\value{
Character vector of engine function names.
}
\description{
List of engine function names exposed by \code{blockr.insurance} to
\code{\link[=new_price_block]{new_price_block()}}'s engine dropdown. Add a new engine here when shipping
a new property pricing variant.
}
\examples{
available_engines()

}
